test(favorite): move delete favorite tests to FavoriteServiceTest
